refactor(client): HASHI-71 예약 상세 컴포넌트 page-local 구조로 정리
