+++ apps/client/src/features/reservation/components/index.ts
@@ -1,20 +1,2 @@
 export { GuestCounter } from './guestCounter'
 export type { GuestCounterProps } from './guestCounter'
-
-export { ReservationProgressSection } from './reservationProgressSection'
-export type {
-  ReservationProgressSectionProps,
-  ReservationProgressStep,
-} from './reservationProgressSection'
-
-export { ReservationReceiptInfoCard } from './reservationReceiptInfoCard'
-export type {
-  ReservationReceiptInfoCardProps,
-  ReservationReceiptInfoItem,
-} from './reservationReceiptInfoCard'
-
-export { ReservationNoticeSection } from './reservationNoticeSection'
-export type { ReservationNoticeSectionProps } from './reservationNoticeSection'
-
-export { ReservationDetailActionBar } from './reservationDetailActionBar'
-export type { ReservationDetailActionBarProps } from './reservationDetailActionBar'
