--- apps/client/src/features/reservation/components/reservationDetailActionBar/index.ts
@@ -1,2 +1,0 @@
-export { ReservationDetailActionBar } from './ReservationDetailActionBar'
-export type { ReservationDetailActionBarProps } from './ReservationDetailActionBar'
