--- apps/client/src/features/reservation/components/reservationNoticeSection/index.ts
@@ -1,2 +1,0 @@
-export { ReservationNoticeSection } from './ReservationNoticeSection'
-export type { ReservationNoticeSectionProps } from './ReservationNoticeSection'
