--- apps/client/src/features/reservation/components/reservationProgressSection/index.ts
@@ -1,6 +1,0 @@
-export { ReservationProgressSection } from './ReservationProgressSection'
-export type {
-  ReservationProgressSectionProps,
-  ReservationProgressStep,
-  ReservationProgressStatus,
-} from './ReservationProgressSection'
