--- apps/client/src/features/reservation/components/reservationReceiptInfoCard/index.ts
@@ -1,5 +1,0 @@
-export { ReservationReceiptInfoCard } from './ReservationReceiptInfoCard'
-export type {
-  ReservationReceiptInfoCardProps,
-  ReservationReceiptInfoItem,
-} from './ReservationReceiptInfoCard'
