--- apps/client/src/features/reservation/constants/index.ts
@@ -1,1 +1,0 @@
-export { reservationNotices } from './reservationNotice'
