+++ apps/client/src/pages/reservationDetail/components/ReservationDetailActionBar.tsx
@@ -10,7 +10,7 @@ export const ReservationDetailActionBar = ({
   onContact,
 }: ReservationDetailActionBarProps) => {
   return (
-    <div className="app-mobile-fixed-bottom bg-white px-7.5 pt-3 pb-[calc(20px+var(--safe-area-bottom,0px))]">
+    <div className="app-mobile-fixed-bottom bg-white px-6 pt-4.25 pb-[calc(20px+var(--safe-area-bottom,0px))]">
       <div className="grid grid-cols-2 gap-2.5">
         <Button onClick={onCancel} size="md" variant="neutral" width="full">
           예약 취소하기
