+++ apps/client/src/pages/reservationDetail/components/ReservationNoticeSection.tsx
