+++ apps/client/src/pages/reservationDetail/components/ReservationReceiptInfoCard.tsx
