+++ apps/client/src/pages/reservationDetail/constants/reservationNotice.ts
